# ElementsService optimization - switch to where-in statement when querying for elements with array of IDs