@brandonkelly 2018-06-02
Fixed this from DbHelper::parseParam() instead, but thanks for the nudge!